+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 
 - Do not fail to edit `Institution` if service to fetch categories is broken.
   [gbastien]
+- Adapted `SelectMeetingWidget` used for the `seances` criterion to make
+  zero count values shown and selectable.
+  [gbastien]
 
 
 1.3.3.1 (2021-08-16)